QuickFit3/TIR-FCCS fit functions:   - bugfixed: "TIR-FCCS: 2D Diffusion+Flow (1/e² radii)"
QuickFit3/imFCCS fit:
  - added example project for directed motion/flow

git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3412 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/imagingfcs/qfrdrimagingfcs_dataeditor_image.cpp
@@ -4370,7 +4370,7 @@ void QFRDRImagingFCSImageEditor::transformImage(double* image, uint32_t width, u
                 }
             } break;
         case QFRDRImagingFCSImageEditor::itSqrt: {
-                for (int32_t i=0; i<width*height; i++) {
+                for (uint64_t i=0; i<width*height; i++) {
                     if (image[i]>=0) image[i]=sqrt(image[i]);
                     else  image[i]=NAN;
                 }
